added some plotting, rearrange imports
